Renaming to match golang conventions better.
