+++ daemon/logger/journaldsemistruct/journald_unsupported.go
