+++ daemon/logger/journaldsemistruct/journaldsemistruct.go
@@ -2,7 +2,7 @@
 
 // Package journald provides the log driver for forwarding server logs
 // to endpoints that receive the systemd format.
-package journald_semistruct
+package journaldsemistruct
 
 import (
 	"fmt"
