+++ daemon/logger/journaldsemistruct/read.go
