+++ daemon/logger/journaldsemistruct/read_native.go
