+++ daemon/logger/journaldsemistruct/read_native_compat.go
