+++ daemon/logger/journaldsemistruct/read_unsupported.go
